Refactoring of common parts between CaseImporter and EventImporter #4041
