+++ sormas-api/src/main/java/de/symeda/sormas/api/importexport/ImportLineResult.java
@@ -12,42 +12,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
+package de.symeda.sormas.api.importexport;
 
-package de.symeda.sormas.backend.event.eventimport;
+public enum ImportLineResult {
 
-public class ImportCellData {
-
-	String value;
-	String entityClass;
-	String[] entityPropertyPath;
-
-	public ImportCellData(String value, String entityClass, String[] entityPropertyPath) {
-		this.value = value;
-		this.entityClass = entityClass;
-		this.entityPropertyPath = entityPropertyPath;
-	}
-
-	public String getValue() {
-		return value;
-	}
-
-	public void setValue(String value) {
-		this.value = value;
-	}
-
-	public String getEntityClass() {
-		return entityClass;
-	}
-
-	public void setEntityClass(String entityClass) {
-		this.entityClass = entityClass;
-	}
-
-	public String[] getEntityPropertyPath() {
-		return entityPropertyPath;
-	}
-
-	public void setEntityPropertyPath(String[] entityPropertyPath) {
-		this.entityPropertyPath = entityPropertyPath;
-	}
+	SUCCESS,
+	ERROR,
+	DUPLICATE;
 }
