+++ sormas-api/src/main/java/de/symeda/sormas/api/importexport/ImportLineResultDto.java
@@ -13,7 +13,7 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package de.symeda.sormas.api.event.eventimport;
+package de.symeda.sormas.api.importexport;
 
 import java.io.Serializable;
