--- sormas-backend/src/main/java/de/symeda/sormas/backend/event/eventimport/ImportErrorException.java
@@ -1,32 +1,0 @@
-/*
- * SORMAS® - Surveillance Outbreak Response Management & Analysis System
- * Copyright © 2016-2020 Helmholtz-Zentrum für Infektionsforschung GmbH (HZI)
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <https://www.gnu.org/licenses/>.
- */
-
-package de.symeda.sormas.backend.event.eventimport;
-
-import de.symeda.sormas.api.i18n.I18nProperties;
-import de.symeda.sormas.api.i18n.Strings;
-
-public class ImportErrorException extends Exception {
-
-	private static final long serialVersionUID = -5852533615013283186L;
-
-	public ImportErrorException(String value, String column) {
-		super(I18nProperties.getString(Strings.errorInvalidValue) + " " + value + " " + I18nProperties.getString(Strings.inColumn) + " " + column);
-	}
-
-	public ImportErrorException(String message) {
-		super(message);
-	}
-}
