+++ sormas-backend/src/main/java/de/symeda/sormas/backend/importexport/ImportCellData.java
@@ -12,11 +12,42 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-package de.symeda.sormas.api.event.eventimport;
 
-public enum ImportLineResult {
+package de.symeda.sormas.backend.importexport;
 
-	SUCCESS,
-	ERROR,
-	DUPLICATE;
+public class ImportCellData {
+
+	String value;
+	String entityClass;
+	String[] entityPropertyPath;
+
+	public ImportCellData(String value, String entityClass, String[] entityPropertyPath) {
+		this.value = value;
+		this.entityClass = entityClass;
+		this.entityPropertyPath = entityPropertyPath;
+	}
+
+	public String getValue() {
+		return value;
+	}
+
+	public void setValue(String value) {
+		this.value = value;
+	}
+
+	public String getEntityClass() {
+		return entityClass;
+	}
+
+	public void setEntityClass(String entityClass) {
+		this.entityClass = entityClass;
+	}
+
+	public String[] getEntityPropertyPath() {
+		return entityPropertyPath;
+	}
+
+	public void setEntityPropertyPath(String[] entityPropertyPath) {
+		this.entityPropertyPath = entityPropertyPath;
+	}
 }
